Add selected state, create abstract bottom sheet dialog class
