+++ commons/src/main/kotlin/com/simplemobiletools/commons/models/SimpleListItem.kt
@@ -4,7 +4,7 @@ import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
 
 @Parcelize
-data class SimpleListItem(val id: Int, val imageRes: Int?, val textRes: Int) : Parcelable {
+data class SimpleListItem(val id: Int, val imageRes: Int?, val textRes: Int, val selected: Boolean = false) : Parcelable {
 
     companion object {
         fun areItemsTheSame(old: SimpleListItem, new: SimpleListItem): Boolean {
@@ -12,7 +12,7 @@ data class SimpleListItem(val id: Int, val imageRes: Int?, val textRes: Int) : P
         }
 
         fun areContentsTheSame(old: SimpleListItem, new: SimpleListItem): Boolean {
-            return old.imageRes == new.imageRes && old.textRes == new.textRes
+            return old.imageRes == new.imageRes && old.textRes == new.textRes && old.selected == new.selected
         }
     }
 }
