+++ commons/src/main/res/drawable/bottom_sheet_bg_black.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/bottom_sheet_background"
+        android:bottom="-2dp">
+        <shape android:shape="rectangle">
+            <solid android:color="@color/md_grey_black" />
+            <stroke
+                android:width="1dp"
+                android:color="@color/light_grey_stroke" />
+            <corners
+                android:topLeftRadius="@dimen/bottom_sheet_corner_radius"
+                android:topRightRadius="@dimen/bottom_sheet_corner_radius" />
+        </shape>
+    </item>
+</layer-list>
