+++ commons/src/main/res/layout/item_simple_list.xml
@@ -23,11 +23,25 @@
         android:id="@+id/bottom_sheet_item_title"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/small_margin"
+        android:layout_marginHorizontal="@dimen/small_margin"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintEnd_toStartOf="@id/bottom_sheet_selected_icon"
         app:layout_constraintStart_toEndOf="@id/bottom_sheet_item_icon"
         app:layout_constraintTop_toTopOf="parent"
         tools:text="ReadMe.md" />
+
+    <androidx.appcompat.widget.AppCompatImageView
+        android:id="@+id/bottom_sheet_selected_icon"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginEnd="@dimen/activity_margin"
+        android:padding="@dimen/medium_margin"
+        android:src="@drawable/ic_check_circle_vector"
+        android:visibility="gone"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toEndOf="@id/bottom_sheet_item_title"
+        app:layout_constraintTop_toTopOf="parent"
+        tools:visibility="visible" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
